jzo2o-ai-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
python-dotenv==1.0.1
nacos-sdk-python==2.0.6
